+++ vkcs/cdn/data_source_shielding_pops.go
@@ -65,7 +65,7 @@ func (d *shieldingPopsDataSource) Read(ctx context.Context, req datasource.ReadR
 
 	tflog.Trace(ctx, "Calling CDN API to list origin shielding POPs")
 
-	shieldingPops, err := shieldingpop.List(client, d.config.GetTenantID()).Extract()
+	shieldingPops, err := shieldingpop.List(client, d.config.GetProjectID()).Extract()
 	if err != nil {
 		resp.Diagnostics.AddError("Error calling CDN API to list origin shielding POP", err.Error())
 		return