+++ vkcs/internal/clients/defaults.go
@@ -1,8 +1,0 @@
-package clients
-
-const (
-	DefaultIdentityEndpoint   = "https://infra.mail.ru/identity/v3/"
-	DefaultUserDomainName     = "users"
-	DefaultRegionName         = "RegionOne"
-	CloudContainersAPIVersion = "1.33"
-)