+++ vkcs/sharedfilesystem/data_source_sharenetwork.go
@@ -92,7 +92,7 @@ func DataSourceSharedFilesystemShareNetwork() *schema.Resource {
 
 func dataSourceSharedFilesystemShareNetworkRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(util.GetRegion(d, config))
+	sfsClient, err := config.SharedFilesystemV2Client(util.GetRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating VKCS sharedfilesystem sfsClient: %s", err)
 	}