+++ vkcs/sharedfilesystem/resource_share_access.go
@@ -85,7 +85,7 @@ func ResourceSharedFilesystemShareAccess() *schema.Resource {
 
 func resourceSharedFilesystemShareAccessCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(util.GetRegion(d, config))
+	sfsClient, err := config.SharedFilesystemV2Client(util.GetRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating VKCS sharedfilesystem client: %s", err)
 	}
@@ -150,7 +150,7 @@ func resourceSharedFilesystemShareAccessCreate(ctx context.Context, d *schema.Re
 
 func resourceSharedFilesystemShareAccessRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(util.GetRegion(d, config))
+	sfsClient, err := config.SharedFilesystemV2Client(util.GetRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating VKCS sharedfilesystem client: %s", err)
 	}
@@ -183,7 +183,7 @@ func resourceSharedFilesystemShareAccessRead(ctx context.Context, d *schema.Reso
 
 func resourceSharedFilesystemShareAccessDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(util.GetRegion(d, config))
+	sfsClient, err := config.SharedFilesystemV2Client(util.GetRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating VKCS sharedfilesystem client: %s", err)
 	}
@@ -249,7 +249,7 @@ func resourceSharedFilesystemShareAccessImport(ctx context.Context, d *schema.Re
 	}
 
 	config := meta.(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(util.GetRegion(d, config))
+	sfsClient, err := config.SharedFilesystemV2Client(util.GetRegion(d, config))
 	if err != nil {
 		return nil, fmt.Errorf("error creating VKCS sharedfilesystem client: %s", err)
 	}