+++ vkcs/sharedfilesystem/resource_share_test.go
@@ -52,7 +52,7 @@ func TestAccSFSShare_basic(t *testing.T) {
 
 func testAccCheckSFSShareDestroy(s *terraform.State) error {
 	config := acctest.AccTestProvider.Meta().(clients.Config)
-	sfsClient, err := config.SharedfilesystemV2Client(acctest.OsRegionName)
+	sfsClient, err := config.SharedFilesystemV2Client(acctest.OsRegionName)
 	if err != nil {
 		return fmt.Errorf("Error creating VKCS sharedfilesystem client: %s", err)
 	}
@@ -83,7 +83,7 @@ func testAccCheckSFSShareExists(n string, share *shares.Share) resource.TestChec
 		}
 
 		config := acctest.AccTestProvider.Meta().(clients.Config)
-		sfsClient, err := config.SharedfilesystemV2Client(acctest.OsRegionName)
+		sfsClient, err := config.SharedFilesystemV2Client(acctest.OsRegionName)
 		if err != nil {
 			return fmt.Errorf("Error creating VKCS sharedfilesystem client: %s", err)
 		}