+++ CHANGELOG.md
@@ -9,6 +9,7 @@ description: |-
 #### v0.14.0 (unreleased)
 - Add vkcs_iam_s3_account resource
 - Add vkcs_iam_s3_account data source
+- Add endpoint overrides and passthrough token authentication
 
 #### v0.13.2
 - Add details to the vkcs_compute_instance creation error